+++ docs/source/conf.py
@@ -406,7 +406,6 @@ nitpick_ignore = [
     ("py:class", "array"),
     ("py:class", "boolean"),
     ("py:class", "null"),
-    ("py:class", "any"),
 ]
