+++ mlflow/entities/document.py
@@ -1,6 +1,6 @@
 from copy import deepcopy
 from dataclasses import asdict, dataclass, field
-from typing import Dict, Optional
+from typing import Any, Dict, Optional
 
 
 @dataclass
@@ -15,13 +15,13 @@ class Document:
     """
 
     page_content: str
-    metadata: Dict[str, any] = field(default_factory=dict)
+    metadata: Dict[str, Any] = field(default_factory=dict)
     id: Optional[str] = None
 
     @classmethod
     def from_langchain_document(cls, document):
         # older versions of langchain do not have the id attribute
-        id = document.id if hasattr(document, "id") else None
+        id = getattr(document, "id", None)
 
         return cls(
             page_content=document.page_content,
@@ -31,13 +31,14 @@ class Document:
 
     @classmethod
     def from_llama_index_node_with_score(cls, node_with_score):
-        metadata = {}
-        metadata["score"] = node_with_score.get_score()
-        # update after setting score so that it can be
-        # overridden if the user wishes to do so
-        metadata.update(deepcopy(node_with_score.metadata))
+        metadata = {
+            "score": node_with_score.get_score(),
+            # update after setting score so that it can be
+            # overridden if the user wishes to do so
+            **deepcopy(node_with_score.metadata),
+        }
 
-        return Document(
+        return cls(
             page_content=node_with_score.get_content(),
             metadata=metadata,
             id=node_with_score.node_id,
