+++ tests/llama_index/test_llama_index_node_conversion.py
@@ -1,15 +1,6 @@
-from langchain_core.documents import Document as LangchainDocument
 from llama_index.core.schema import NodeWithScore, TextNode
 
 from mlflow.entities import Document
-
-
-def test_from_langchain_document():
-    langchain_document = LangchainDocument(page_content="Hello", metadata={"key": "value"})
-    document = Document.from_langchain_document(langchain_document)
-    assert document.page_content == "Hello"
-    assert document.metadata == {"key": "value"}
-    assert document.id is None
 
 
 def test_from_llama_index_node_with_score():
